+++ src/test/kotlin/io/kjson/pointer/ExtensionTest.kt
@@ -27,6 +27,7 @@ package io.kjson.pointer
 
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
+import kotlin.test.assertFalse
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 import kotlin.test.assertSame
@@ -40,10 +41,10 @@ import io.kjson.JSON
 import io.kjson.JSON.asInt
 import io.kjson.JSONArray
 import io.kjson.JSONBoolean
-import io.kjson.JSONIncorrectTypeException
 import io.kjson.JSONInt
 import io.kjson.JSONObject
 import io.kjson.JSONString
+import io.kjson.JSONTypeException
 import io.kjson.pointer.test.SampleJSON.testArray
 import io.kjson.pointer.test.SampleJSON.testMixedArray
 import io.kjson.pointer.test.SampleJSON.testObject
@@ -65,6 +66,90 @@ class ExtensionTest {
         }
     }
 
+    @Test fun `should map contents of array to its primitive type`() {
+        val json = JSON.parseArray("[1,10,100,1000,10000]")
+        val ref = JSONRef(json)
+        val array = ref.map<JSONInt, Int>()
+        expect(listOf(1, 10, 100, 1000, 10000)) { array }
+    }
+
+    @Test fun `should throw exception when 'map' operation finds invalid value`() {
+        val json = JSON.parseArray("""[1,10,"error",1000,10000]""")
+        val ref = JSONRef(json)
+        assertFailsWith<JSONTypeException> { ref.map<JSONInt, Int> { value * (it + 1) } }.let {
+            expect("Child not correct type (JSONInt), was \"error\", at /2") { it.message }
+            expect("Child") { it.nodeName }
+            expect("JSONInt") { it.target }
+            expect(JSONString("error")) { it.value }
+            expect(JSONPointer("/2")) { it.key }
+        }
+    }
+
+    @Test fun `should map contents of array with transformation`() {
+        val json = JSON.parseArray("[1,10,100,1000,10000]")
+        val ref = JSONRef(json)
+        val array = ref.map<JSONInt, Int> { value * (it + 1) }
+        expect(listOf(1, 20, 300, 4000, 50000)) { array }
+    }
+
+    @Test fun `should throw exception when transforming 'map' operation finds invalid value`() {
+        val json = JSON.parseArray("[1,10,100,false,10000]")
+        val ref = JSONRef(json)
+        assertFailsWith<JSONTypeException> { ref.map<JSONInt, Int> { value * (it + 1) } }.let {
+            expect("Child not correct type (JSONInt), was false, at /3") { it.message }
+            expect("Child") { it.nodeName }
+            expect("JSONInt") { it.target }
+            expect(JSONBoolean.FALSE) { it.value }
+            expect(JSONPointer("/3")) { it.key }
+        }
+    }
+
+    @Test fun `should map contents of array with complex transformation`() {
+        val json = JSON.parseArray("""[{"f1":"AAA"},{"f1":"BBB"},{"f1":"CCC"},{"f1":"DDD"}]""")
+        val ref = JSONRef(json)
+        val array = ref.map<JSONObject, String> { child<JSONString>("f1").value }
+        expect(listOf("AAA", "BBB", "CCC", "DDD")) { array }
+    }
+
+    @Test fun `should test whether any array item matches predicate`() {
+        val json = JSON.parseArray("[1,2,3,4,5]")
+        val ref = JSONRef(json)
+        assertTrue { ref.any<JSONInt> { value > 3 } }
+        assertFalse { ref.any<JSONInt> { value > 10 } }
+    }
+
+    @Test fun `should throw exception when 'any' test finds invalid value`() {
+        val json = JSON.parseArray("""[1,2,3,"wrong",5]""")
+        val ref = JSONRef(json)
+        assertFailsWith<JSONTypeException> { ref.any<JSONInt> { value > 3 } }.let {
+            expect("Child not correct type (JSONInt), was \"wrong\", at /3") { it.message }
+            expect("Child") { it.nodeName }
+            expect("JSONInt") { it.target }
+            expect(JSONString("wrong")) { it.value }
+            expect(JSONPointer("/3")) { it.key }
+        }
+    }
+
+    @Test fun `should test whether all array items match predicate`() {
+        val json = JSON.parseArray("[1,2,3,4,5]")
+        val ref = JSONRef(json)
+        assertTrue { ref.all<JSONInt> { value < 10 } }
+        assertFalse { ref.all<JSONInt> { value < 5 } }
+    }
+
+    @Test fun `should throw exception when 'all' test finds invalid value`() {
+        val json = JSON.parseArray("""[1,2,"bad",4,5]""")
+        val ref = JSONRef(json)
+        assertFailsWith<JSONTypeException> { ref.all<JSONInt> { value < 10 } }.let {
+            expect("Child not correct type (JSONInt), was \"bad\", at /2") { it.message }
+            expect("Child") { it.nodeName }
+            expect("JSONInt") { it.target }
+            expect(JSONString("bad")) { it.value }
+            expect(JSONPointer("/2")) { it.key }
+        }
+    }
+
+    @Suppress("DEPRECATION")
     @Test fun `should unconditionally map value`() {
         val json = JSON.parseObject("""{"a":1,"b":1,"c":2,"d":3,"e":5}""")
         val ref = JSONRef(json)
@@ -75,12 +160,21 @@ class ExtensionTest {
         assertFailsWith<JSONPointerException> {
             ref.map<JSONInt, Int>("x") { it.value }
         }.let {
-            expect("Node does not exist - \"/x\"") { it.message }
+            expect("Node does not exist, at /x") { it.message }
         }
         val e: Int = ref.map("e") { prop: JSONInt -> prop.value }
         expect(5) { e }
+        // alternative, following deprecation of original function, using recommended replacement code
+        val dd = ref.child<JSONInt>("d").value
+        expect(3) { dd }
+        assertFailsWith<JSONPointerException> {
+            ref.child<JSONInt>("x").value
+        }.let {
+            expect("Node does not exist, at /x") { it.message }
+        }
     }
 
+    @Suppress("DEPRECATION")
     @Test fun `should conditionally map value`() {
         val json = JSON.parseObject("""{"a":1,"b":1,"c":2,"d":3,"e":5}""")
         val ref = JSONRef(json)
@@ -94,6 +188,11 @@ class ExtensionTest {
         assertNull(x)
         val e: Int? = ref.mapIfPresent("e") { prop: JSONInt -> prop.value }
         expect(5) { e }
+        // alternative, following deprecation of original function, using recommended replacement code
+        val dd = ref.optionalChild<JSONInt>("d")?.value
+        expect(3) { dd }
+        val xx = ref.optionalChild<JSONInt>("x")?.value
+        assertNull(xx)
     }
 
     @Test fun `should get optional String`() {
@@ -102,7 +201,7 @@ class ExtensionTest {
         expect("Hello") { ref.optionalString("aaa") }
         expect("World") { ref.optionalString("bbb") }
         assertNull(ref.optionalString("ccc"))
-        assertFailsWith<JSONIncorrectTypeException> { ref.optionalString("xxx") }.let {
+        assertFailsWith<JSONTypeException> { ref.optionalString("xxx") }.let {
             expect("Node not correct type (String), was 999, at /xxx") { it.message }
             expect("Node") { it.nodeName }
             expect("String") { it.target }
@@ -117,7 +216,7 @@ class ExtensionTest {
         expect(true) { ref.optionalBoolean("aaa") }
         expect(false) { ref.optionalBoolean("bbb") }
         assertNull(ref.optionalBoolean("ccc"))
-        assertFailsWith<JSONIncorrectTypeException> { ref.optionalBoolean("xxx") }.let {
+        assertFailsWith<JSONTypeException> { ref.optionalBoolean("xxx") }.let {
             expect("Node not correct type (Boolean), was 999, at /xxx") { it.message }
             expect("Node") { it.nodeName }
             expect("Boolean") { it.target }
@@ -132,7 +231,7 @@ class ExtensionTest {
         expect(123) { ref.optionalInt("aaa") }
         expect(456) { ref.optionalInt("bbb") }
         assertNull(ref.optionalInt("ccc"))
-        assertFailsWith<JSONIncorrectTypeException> { ref.optionalInt("xxx") }.let {
+        assertFailsWith<JSONTypeException> { ref.optionalInt("xxx") }.let {
             expect("Node not correct type (Int), was \"wrong\", at /xxx") { it.message }
             expect("Node") { it.nodeName }
             expect("Int") { it.target }
@@ -147,7 +246,7 @@ class ExtensionTest {
         expect(123) { ref.optionalLong("aaa") }
         expect(123456789123456789) { ref.optionalLong("bbb") }
         assertNull(ref.optionalLong("ccc"))
-        assertFailsWith<JSONIncorrectTypeException> { ref.optionalLong("xxx") }.let {
+        assertFailsWith<JSONTypeException> { ref.optionalLong("xxx") }.let {
             expect("Node not correct type (Long), was \"wrong\", at /xxx") { it.message }
             expect("Node") { it.nodeName }
             expect("Long") { it.target }
@@ -163,7 +262,7 @@ class ExtensionTest {
         expect(BigDecimal(123456789123456789)) { ref.optionalDecimal("bbb") }
         expect(BigDecimal("1.5")) { ref.optionalDecimal("ccc") }
         assertNull(ref.optionalDecimal("ddd"))
-        assertFailsWith<JSONIncorrectTypeException> { ref.optionalDecimal("xxx") }.let {
+        assertFailsWith<JSONTypeException> { ref.optionalDecimal("xxx") }.let {
             expect("Node not correct type (Decimal), was \"wrong\", at /xxx") { it.message }
             expect("Node") { it.nodeName }
             expect("Decimal") { it.target }
@@ -179,7 +278,7 @@ class ExtensionTest {
         assertNotNull(childRef)
         expect(1000) { childRef.node["bbb"].asInt }
         assertNull(ref.optionalChild<JSONObject>("ccc"))
-        assertFailsWith<JSONIncorrectTypeException> { ref.optionalChild<JSONObject>("xxx") }.let {
+        assertFailsWith<JSONTypeException> { ref.optionalChild<JSONObject>("xxx") }.let {
             expect("Child not correct type (JSONObject), was 999, at /xxx") { it.message }
             expect("Child") { it.nodeName }
             expect("JSONObject") { it.target }
@@ -195,7 +294,7 @@ class ExtensionTest {
         assertNotNull(childRef)
         expect(888) { childRef.node[0].asInt }
         assertNull(ref.optionalChild<JSONArray>("ccc"))
-        assertFailsWith<JSONIncorrectTypeException> { ref.optionalChild<JSONArray>("xxx") }.let {
+        assertFailsWith<JSONTypeException> { ref.optionalChild<JSONArray>("xxx") }.let {
             expect("Child not correct type (JSONArray), was 999, at /xxx") { it.message }
             expect("Child") { it.nodeName }
             expect("JSONArray") { it.target }
@@ -248,6 +347,7 @@ class ExtensionTest {
 
     @Test fun `should create a reference to a JSONObject`() {
         val ref = testObject.ref()
+        assertTrue(ref.isRef<JSONObject>())
         assertSame(ref.base, testObject)
         assertSame(ref.node, testObject)
         expect(JSONPointer.root) { ref.pointer }
@@ -255,6 +355,7 @@ class ExtensionTest {
 
     @Test fun `should create a reference to a JSONArray`() {
         val ref = testArray.ref()
+        assertTrue(ref.isRef<JSONArray>())
         assertSame(ref.base, testArray)
         assertSame(ref.node, testArray)
         expect(JSONPointer.root) { ref.pointer }